Clean up and update files around new query
