+++ frontend/src/metabase/new_dataset/components/NewDatasetOption.jsx
@@ -4,7 +4,7 @@ import cx from "classnames";
 import { Link } from "react-router";
 import { color } from "metabase/lib/colors";
 
-export default class NewQueryOption extends Component {
+export default class NewDatasetOption extends Component {
   state = {
     hover: false,
   };
